requirements.txt
py_assimilator==1.1.1
pydantic==1.10.5
typing_extensions==4.5.0
